Move firebase configs from index.js to /firebase.js
